# CodingChallengeAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CodingChallengeAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Comentario/ComentariosByIdFilme": {
      "get": {
        "tags": [
          "Comentario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idFilme",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Comentario": {
      "post": {
        "tags": [
          "Comentario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Comentario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Comentario/AvaliarComentario": {
      "post": {
        "tags": [
          "Comentario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ComentarioAvaliacao"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Comentario/MarcarComoRepetido": {
      "put": {
        "tags": [
          "Comentario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idComentario",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "idUsuario",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "repetido",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Comentario/ExcluirComentario": {
      "delete": {
        "tags": [
          "Comentario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idComentario",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "idUsuario",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Filme/FilmesApi": {
      "get": {
        "tags": [
          "Filme"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "busca",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Filme/FilmesApiPorId": {
      "get": {
        "tags": [
          "Filme"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idFilme",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Nota/NotasByFIlme": {
      "get": {
        "tags": [
          "Nota"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idFilme",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Nota": {
      "post": {
        "tags": [
          "Nota"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Nota"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Resposta/RespostasByComentario": {
      "get": {
        "tags": [
          "Resposta"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idComentario",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Resposta": {
      "post": {
        "tags": [
          "Resposta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Resposta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Usuario/Usuarios": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Usuario/Usuario": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idUsuario",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Usuario/Cadastrar": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Usuario/TornarModerador": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idUsuarioModerador",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "idUsuario",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Comentario": {
      "type": "object",
      "properties": {
        "idComentario": {
          "format": "int32",
          "type": "integer"
        },
        "idFilme": {
          "type": "string"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "texto": {
          "type": "string"
        },
        "gostei": {
          "format": "int32",
          "type": "integer"
        },
        "naoGostei": {
          "format": "int32",
          "type": "integer"
        },
        "repetido": {
          "type": "boolean"
        }
      }
    },
    "ComentarioAvaliacao": {
      "type": "object",
      "properties": {
        "idAvaliacao": {
          "format": "int32",
          "type": "integer"
        },
        "idComentario": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "gostei": {
          "type": "boolean"
        }
      }
    },
    "Nota": {
      "type": "object",
      "properties": {
        "idNota": {
          "format": "int32",
          "type": "integer"
        },
        "idFilme": {
          "type": "string"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "valorNota": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "PerfilUsuario": {
      "format": "int32",
      "enum": [
        1,
        2,
        3,
        4
      ],
      "type": "integer"
    },
    "Resposta": {
      "type": "object",
      "properties": {
        "idResposta": {
          "format": "int32",
          "type": "integer"
        },
        "idComentario": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "texto": {
          "type": "string"
        }
      }
    },
    "Usuario": {
      "type": "object",
      "properties": {
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        },
        "pontos": {
          "format": "int32",
          "type": "integer"
        },
        "perfil": {
          "$ref": "#/definitions/PerfilUsuario"
        }
      }
    }
  }
}